refactoring to put styles for larger components into a style file
